+++ dlls/game/mp_modeTeamBase.cpp
@@ -110,7 +110,7 @@ bool ModeTeamBase::isEndOfMatch( void )
 	for (int idx=1; idx <= _teamList.NumObjects(); idx++) 
 	{
 		Team* team = _teamList.ObjectAt(idx);
-		if (team->getDeaths() > getPointLimit())
+		if ((int)(team->getDeaths()) > getPointLimit())
 		{
 			multiplayerManager.centerPrintAllClients(va("$$%s$$ $$TeamLoses$$\n", team->getName().c_str() ), CENTERPRINT_IMPORTANCE_NORMAL );
