+++ dlls/game/navigate.h
@@ -714,9 +714,9 @@ void PathFinder<Heuristic>::PropagateDown
 		child = thePathManager.GetNode( path->targetNodeIndex );
 
 		movecost = g + heuristic.cost( node, c );
-		if ( movecost < child->g )
+		if ( (long)movecost < child->g )
 			{
-			child->g = movecost;
+			child->g = (long)movecost;
 			child->f = child->g + child->h;
 			child->Parent = node;
 
@@ -739,9 +739,9 @@ void PathFinder<Heuristic>::PropagateDown
 			// we stop the propagation when the g value of the child is equal or better than
 			// the cost we're propagating
 			movecost = parent->g + path->moveCost;
-			if ( movecost < child->g )
+			if ( (long)movecost < child->g )
     			{
-				child->g = movecost;
+				child->g = (long)movecost;
 				child->f = child->g + child->h;
     			child->Parent = parent;
     			stack.Push( child );
